Fixed: First wave of linting

+++ common/helpers.go
@@ -73,7 +73,7 @@ func BuildNfgenMsg(family int, version uint8, resID uint16, n *syscall.NlMsghdr)
 	n.Len = NlMsgLength(SizeofNfGenMsg)
 	return &NfqGenMsg{
 		nfgenFamily: uint8(family),
-		version:     uint8(version),
+		version:     version,
 		resID:       uint16(resID),
 	}
 }
